+++ cmd/tfproviderlint/tfproviderlint.go
@@ -41,6 +41,7 @@ import (
 	"github.com/bflad/tfproviderlint/passes/S018"
 	"github.com/bflad/tfproviderlint/passes/S019"
 	"github.com/bflad/tfproviderlint/passes/S020"
+	"github.com/bflad/tfproviderlint/passes/V001"
 )
 
 func main() {
@@ -79,5 +80,6 @@ func main() {
 		S018.Analyzer,
 		S019.Analyzer,
 		S020.Analyzer,
+		V001.Analyzer,
 	)
 }